Revert "Feature/elelog 28 endpoint for reverting document to repository version (#1499)"
This reverts commit 014d31fd0af4207daeb394a4333ab32f0fa67dfc.
