+++ src-srv/api/documents/addassignment.ts
@@ -1,33 +1,36 @@
 import type { Request } from 'express'
-import type { RouteContentResponse, RouteHandler, RouteStatusResponse } from '../../../routes.js'
-import logger from '../../../lib/logger.js'
+import type { RouteContentResponse, RouteHandler, RouteStatusResponse } from '../../routes.js'
+import logger from '../../lib/logger.js'
 import { Block } from '@ttab/elephant-api/newsdoc'
-import { appendAssignment, appendDocumentToAssignment } from '../../../lib/createYItem.js'
-import { createSnapshot } from '../../../utils/createSnapshot.js'
-import { planningDocumentTemplate } from '../../../../shared/templates/planningDocumentTemplate.js'
-import { getDeliverableType } from '../../../../src/defaults/templates/lib/getDeliverableType.js'
+import { appendAssignment, appendDocumentToAssignment } from '../../lib/createYItem.js'
+import type { Context } from '../../lib/assertContext.js'
+import { assertContext } from '../../lib/assertContext.js'
+import { createSnapshot } from '../../utils/createSnapshot.js'
+import { planningDocumentTemplate } from '../../../shared/templates/planningDocumentTemplate.js'
+import { getDeliverableType } from '../../../src/defaults/templates/lib/getDeliverableType.js'
 import { toYjsNewsDoc } from '@/shared/transformations/yjsNewsDoc.js'
 import { toGroupedNewsDoc } from '@/shared/transformations/groupedNewsDoc.js'
-import type { Wire } from '../../../../src/hooks/index/useDocuments/schemas/wire.js'
-import { getContextFromValidSession, isContext, type Context } from '../../../lib/context.js'
+import type { Wire } from '../../../src/hooks/index/useDocuments/schemas/wire.js'
 
 type Response = RouteContentResponse | RouteStatusResponse
 
 /**
- * Add assignment to an existing planning or a newly created one.
+ * Add assignment to an existing planning or a newly craeted one.
  */
 export const POST: RouteHandler = async (req: Request, { collaborationServer, res }) => {
-  const planningId = req.params.id
   const locals = res.locals as Record<string, unknown> | undefined
   const session = locals?.session as { accessToken?: string, user?: Context['user'] } | undefined
 
-  const context = getContextFromValidSession(session)
-  if (!isContext(context)) {
-    return context
+  if (!session?.accessToken || !session?.user) {
+    return {
+      statusCode: 401,
+      statusMessage: 'Unauthorized: Session not found, can not snapshot document'
+    }
   }
 
   // Validate incoming data
   const {
+    planningId,
     planningTitle,
     type,
     deliverableId,
@@ -70,6 +73,19 @@ export const POST: RouteHandler = async (req: Request, { collaborationServer, re
     return {
       statusCode: 400,
       statusMessage: 'Invalid input to document addassignment endpoint'
+    }
+  }
+
+  const context: Context = {
+    accessToken: session.accessToken,
+    user: session.user,
+    agent: 'server'
+  }
+
+  if (!assertContext(context)) {
+    return {
+      statusCode: 500,
+      statusMessage: 'Invalid context provided'
     }
   }
