+++ src-srv/api/snapshot/[id]/index.ts
@@ -1,9 +1,10 @@
 import logger from '../../../lib/logger.js'
 import type { Request } from 'express'
 import type { RouteContentResponse, RouteHandler, RouteStatusResponse } from '../../../routes.js'
+import type { Context } from '../../../lib/assertContext.js'
+import { assertContext } from '../../../lib/assertContext.js'
 import { createSnapshot } from '../../../utils/createSnapshot.js'
 import type { Session } from 'next-auth'
-import { getContextFromValidSession, isContext, type Context } from '../../../lib/context.js'
 
 type Response = RouteContentResponse | RouteStatusResponse
 
@@ -30,9 +31,24 @@ export const GET: RouteHandler = async (req: Request, { collaborationServer, cac
     }
   }
 
-  const context = getContextFromValidSession(session)
-  if (!isContext(context)) {
-    return context
+  if (!accessToken || !user) {
+    return {
+      statusCode: 401,
+      statusMessage: 'Unauthorized: Session not found, can not snapshot document'
+    }
+  }
+
+  const context: Context = {
+    accessToken,
+    user,
+    agent: 'server'
+  }
+
+  if (!assertContext(context)) {
+    return {
+      statusCode: 500,
+      statusMessage: 'Invalid context provided'
+    }
   }
 
   try {
